pyproject.toml
[tool.poetry]
name = "ta-chatpdf"
version = "0.0.1"
description = ""
authors = ["Fikry C Farenza <vickyfarenza@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0.post1"
langchain = "^0.2.11"
langchain-community = "^0.2.10"
langchain-openai = "^0.1.17"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
